requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
asyncpg==0.29.0
apscheduler==3.10.4
httpx==0.27.2
pytz==2024.2